requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pymongo==4.5.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.9
python-dotenv==1.0.0

